+++ lib/slkecho/options.rb
@@ -2,6 +2,6 @@
 
 module Slkecho
   class Options
-    attr_accessor :channel, :mention, :message, :username, :icon_url, :icon_emoji
+    attr_accessor :channel, :mention_by_email, :message, :username, :icon_url, :icon_emoji
   end
 end