+++ lib/slkecho/slack_request/post_message.rb
@@ -3,7 +3,7 @@
 module Slkecho
   module SlackRequest
     class PostMessage
-      Params = Struct.new(:channel, :message, :user_id, :username, :icon_url, :icon_emoji, keyword_init: true)
+      Params = Struct.new(:channel, :message, :username, :icon_url, :icon_emoji, keyword_init: true)
 
       def initialize(slack_api_token:)
         @slack_api_token = slack_api_token
@@ -34,15 +34,23 @@ module Slkecho
       def request_body(params)
         {
           "channel" => params.channel,
-          "text" => text_from(params.message, params.user_id),
+          "blocks" => blocks_from(params.message),
           "username" => params.username,
           "icon_url" => params.icon_url,
           "icon_emoji" => params.icon_emoji
         }
       end
 
-      def text_from(message, user_id = nil)
-        user_id.nil? ? message : "<@#{user_id}> #{message}"
+      def blocks_from(message)
+        [
+          {
+            "type" => "section",
+            "text" => {
+              "type" => "mrkdwn",
+              "text" => message
+            }
+          }
+        ]
       end
     end
   end