+++ spec/lib/slkecho/option_parser_spec.rb
@@ -7,25 +7,28 @@ RSpec.describe Slkecho::OptionParser do
     context "when no arguments" do
       let(:argv) { %w[] }
 
-      it { is_expected.to have_attributes(channel: nil, mention: nil, message: nil) }
+      it { is_expected.to have_attributes(channel: nil, mention_by_email: nil, message: nil) }
     end
 
     context "when channel option is given" do
       let(:argv) { %w[-c #general] }
 
-      it { is_expected.to have_attributes(channel: "#general", mention: nil, message: nil) }
+      it { is_expected.to have_attributes(channel: "#general", mention_by_email: nil, message: nil) }
     end
 
     context "when channel and message option is given" do
       let(:argv) { %w[-c #general message] }
 
-      it { is_expected.to have_attributes(channel: "#general", mention: nil, message: "message") }
+      it { is_expected.to have_attributes(channel: "#general", mention_by_email: nil, message: "message") }
     end
 
     context "when channel and message and to option is given" do
       let(:argv) { %w[-c #general -m user1@example.com message] }
 
-      it { is_expected.to have_attributes(channel: "#general", mention: "user1@example.com", message: "message") }
+      it {
+        expect(subject).to have_attributes(channel: "#general", mention_by_email: "user1@example.com",
+                                           message: "message")
+      }
     end
 
     context "when message is from stdin" do
@@ -35,7 +38,7 @@ RSpec.describe Slkecho::OptionParser do
       before { $stdin = stdin }
       after { $stdin = STDIN }
 
-      it { is_expected.to have_attributes(channel: "#general", mention: nil, message: "message") }
+      it { is_expected.to have_attributes(channel: "#general", mention_by_email: nil, message: "message") }
     end
 
     context "when message is nothing" do